requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple

dynaconf==3.2.6
fastapi==0.115.4
loguru==0.7.2
milvus_cli==1.0.0
pymilvus[model]==2.4.3
pydantic==2.9.2
requests==2.31.0
tqdm==4.66.6
uvicorn==0.32.0
torch>=2.6.0

# 修复 marshmallow 版本冲突（pymilvus 依赖 environs，需要 marshmallow < 3.22）
marshmallow==3.21.3

# Embedding 模型依赖
FlagEmbedding>=1.3.3
datasets
peft

# 文档解析
pdfplumber==0.11.0
python-docx==1.1.0

# 文件上传支持
python-multipart==0.0.9

# 爬虫相关
selenium==4.15.2